backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
google-genai
slowapi
sentry-sdk
pypdf
python-docx
